backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
# Use a newer SDK that supports gemini-2.5-flash
google-generativeai>=0.8.0,<1.0
python-dotenv==1.0.0
